requirements.txt
certifi==2018.10.15
chardet==3.0.4
ChatterBot==0.8.7
chatterbot-corpus==1.1.4
Click==7.0
Flask==1.0.2
future==0.17.1
idna==2.7
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.0
mathparse==0.1.2
nltk==3.3
oauthlib==2.1.0
pymongo==3.7.2
python-dateutil==2.6.1
python-twitter==3.4.2
PyYAML==3.13
requests==2.20.0
requests-oauthlib==1.0.0
six==1.11.0
SQLAlchemy==1.2.13
urllib3==1.24
Werkzeug==0.14.1
